pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sgmy-food-recognition"
version = "0.1.0"
description = "Singapore & Malaysian Food Recognition using Qwen2.5-VL"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "hong85"}
]
keywords = [
    "food-recognition",
    "computer-vision",
    "singapore",
    "malaysia",
    "qwen",
    "vlm",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.49.0",
    "accelerate>=0.26.0",
    "pillow",
    "requests",
    "qwen-vl-utils",
]

[project.optional-dependencies]
training = [
    "peft>=0.10.0",
    "trl>=0.8.6",
    "bitsandbytes>=0.43.0",
    "datasets>=2.18.0",
]
dataset = [
    "img2dataset",
    "pandas",
    "pyarrow",
    "duckduckgo_search",
]
hub = [
    "huggingface_hub>=0.20.0",
]
viz = [
    "matplotlib",
]
all = [
    "sgmy-food-recognition[training,dataset,hub,viz]",
]
dev = [
    "pytest",
    "black",
    "ruff",
]

[project.urls]
Homepage = "https://github.com/ghtliew18/sg-my-food-recognition"
Repository = "https://github.com/ghtliew18/sg-my-food-recognition"

[project.scripts]
sgmy-food = "sgmy_food.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["sgmy_food*"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "W"]
